composer/app/requirements.txt
Flask
mysql-connector
SQLAlchemy
mysqlclient
requests
